requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
aiosqlite==0.20.0
httpx==0.28.1
Pillow==11.0.0
opencv-python-headless==4.10.0.84
numpy==1.26.4
scipy==1.14.1
pydantic==2.10.3
python-multipart==0.0.20
sse-starlette==2.1.3
pytest==8.3.4
pytest-asyncio==0.25.2
